fix(stock): handling delete stock

+++ app/Http/Controllers/StockController.php
@@ -38,21 +38,23 @@ class StockController extends Controller
             'stock_name' => 'required',
             'stock_quantity' => 'required|numeric',
             "stock_satuan" => "required",
-            "stock_description" => "required",
-            'stock_photo' => 'required|image|mimes:jpeg,png,jpg|max:1024',
+            'stock_photo' => 'image|mimes:jpeg,png,jpg|max:1024',
         ]);
 
         // get file
         $file = request()->file('stock_photo');
-        $file_name = md5($file->getClientOriginalName() . time()) . '.' . $file->getClientOriginalExtension();
-        $file->move('storage/stocks', $file_name);
+
+        if ($file) {
+            $file_name = md5($file->getClientOriginalName() . time()) . '.' . $file->getClientOriginalExtension();
+            $file->move('storage/stocks', $file_name);
+        }
 
         Stock::create([
             'stock_name' => request('stock_name'),
             'stock_quantity' => request('stock_quantity'),
             'stock_satuan' => request('stock_satuan'),
-            'stock_description' => request('stock_description'),
-            'stock_photo' => $file_name,
+            'stock_description' => request('stock_description') ?? "",
+            'stock_photo' => $file_name ?? "",
         ]);
         return back()->with('stock.success', 'Stock berhasil ditambahkan.');
     }
@@ -107,7 +109,10 @@ class StockController extends Controller
 
             // unlink the file
             $stock = Stock::where("stock_id", $id)->first();
-            unlink(public_path('storage/stocks/' . $stock->stock_photo));
+
+            if ($stock->stock_photo && file_exists(public_path('storage/stocks/' . $stock->stock_photo))) {
+                unlink(public_path('storage/stocks/' . $stock->stock_photo));
+            }
 
             Stock::where("stock_id", $id)->delete();
 
@@ -118,7 +123,7 @@ class StockController extends Controller
         } catch (\Throwable $th) {
             return response()->json([
                 "status" => false,
-                "message" => "Stock gagal dihapus.",
+                "message" => $th->getMessage() ?? "Stock gagal dihapus",
             ])->setStatusCode(500);
         }
     }
